# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenAPI",
        "description": "L5 Swagger OpenAPI description",
        "contact": {
            "email": "darius@matulion.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/feedback/{id_event}": {
            "get": {
                "tags": [
                    "feedback"
                ],
                "summary": "Lấy ra tất cả phản hồi từ người dùng gửi đến theo sự kiện",
                "description": "\n     *      - Endpoint này cho phép lấy ra thông tin phản hồi theo sự kiện.\n     *      - Trả về thông tin của các user được cài đặt.\n     *      - Role được sử dụng quản quản lí, nhân viên, sinh viên\n     *      - id_event là id sự kiện ",
                "operationId": "9d2f400cbd177f419d6d94af06142ae4",
                "responses": {
                    "200": {
                        "description": "Dữ liệu trả về thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy tất cả phản hồi thành công"
                                        },
                                        "metadata": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "hậu đặng vippro max"
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "event_id": {
                                                        "type": "integer",
                                                        "example": "2"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-11-28 17:02:29"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-11-28 17:02:29"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": "1"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Kurtis Legros IV"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "haudvph20519@fpt.edu.vn"
                                                            },
                                                            "phone": {
                                                                "type": "string",
                                                                "example": "+1 (564) 267-3494"
                                                            },
                                                            "role": {
                                                                "type": "integer",
                                                                "example": "1"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2023-12-02T08:55:45.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2023-12-02T08:55:45.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}